[#16] refactor : ParticipatePetGroupModal 컴포넌트 분리
